requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
pyTelegramBotAPI==4.4.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
transitions==0.8.11
urllib3==1.26.9
